% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astar.R
\name{astar}
\alias{astar}
\title{Runs the A* algorithm.}
\usage{
astar(start, goal, params, max.iters = Inf)
}
\arguments{
\item{start}{initial node in the search space}

\item{goal}{final node in the search space}

\item{params}{a list containing functions necessary to define the search space}

\item{max.iters}{maximum number of iterations to run - for debuggin purposes only, defaults to \code{Inf}}
}
\value{
A list containing two components \code{solution} and \code{history}.
}
\description{
The search space is defined by a list passed thorugh params.
}
\examples{

# Searching for a path from (0,0), to (0,3) with three points blocked along the way

dirs <- list(c(1,0),c(0,1),c(0,-1),c(-1,0))
params <- list(
 heuristic = function(el, goal) sum((goal-el)^2),
 distance = function(el, parent, parentDistance) parentDistance + 1,
 neighbours = function(el) setdiff(lapply(dirs, function(d) el+d), list(c(0,2), c(1,1), c(-1,1)))
 )

 res <- astar(c(0,0), c(0,3), params)

}

